# Add explicit EOF for expression parser and use assert instead of exception in sql planner