--- a/sql/src/main/java/org/apache/druid/sql/calcite/filtration/BottomUpTransform.java
+++ b/sql/src/main/java/org/apache/druid/sql/calcite/filtration/BottomUpTransform.java
@@ -36,7 +36,8 @@ public abstract class BottomUpTransform implements Function<Filtration, Filtrati
   private DimFilter checkedProcess(final DimFilter filter)
   {
     final DimFilter retVal = process(Preconditions.checkNotNull(filter, "filter"));
-    return Preconditions.checkNotNull(retVal, "process(filter) result in %s", getClass().getSimpleName());
+    assert retVal != null;
+    return retVal;
   }
 
   @Override